[chore] MessageItem 파일 분리
