+++ NearTalk/NearTalk/Presentation/Chat/ViewModel/ChatViewModel.swift
@@ -243,26 +243,3 @@ private extension DefaultChatViewModel {
         }
     }
 }
-
-struct MessageItem: Hashable {
-    var id: String // message uuid
-    var userName: String // sender name
-    var message: String?
-    var type: MyMessageType
-    var imagePath: String?
-    var createdDate: Date
-    
-    init(chatMessage: ChatMessage, myID: String, userProfile: UserProfile?) {
-        self.id = chatMessage.uuid ?? UUID().uuidString
-        self.userName = userProfile?.username  ?? "알수없음"
-        self.message = chatMessage.text
-        self.type = chatMessage.senderID == myID ? MyMessageType.send : MyMessageType.receive
-        self.imagePath = userProfile?.profileImagePath ?? "이미지없음"
-        self.createdDate = chatMessage.createdDate ?? Date()
-    }
-}
-
-enum MyMessageType: String {
-    case send
-    case receive
-}
